Move sequencer and L1 chain code to separate files
