+++ runner/network/fault_proof_benchmark.go
@@ -1,0 +1,1 @@
+package network
